Import the BupBoop library, needed for "Rikki & Vikki".
This is located at: http://tailchao.com/Audio/index.php#BupBoop

It uses the zlib license, which is GPL compatible.

This library is needed to emulate the sound hardware that "Rikki &
Vikki" uses. It's an entirely separate chip included in the cartridge
that produces 16-bit stereo audio.

Note that BupBoop supplies a library called WinTone that provides
DirectSound bindings, but we don't use it both because we already use
DirectSound and because using WinTone would complicate `libretro`
portability.